airflow/requirements.txt
apache-airflow-providers-google
pyarrow
pandas
fastparquet